MAGE-973 Reorganize unit tests
